third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-style-element/style_type_svg-expected.txt

This is a testharness.js-based test.
[FAIL] With a mixed-case type attribute, the style should apply
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
